Refactor work-order processing to an abstract class
